% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishproject-package.R
\docType{package}
\name{fishproject-package}
\alias{fishproject}
\alias{fishproject-package}
\title{fishproject: Set up a standardised R project}
\description{
This package allows you to automatically create a standard project structure, potentially using stan and/or drake
}
\author{
\strong{Maintainer}: Nina Schiettekatte \email{nina.schiettekatte@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
